requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bootstrap4==0.1.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.3
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap4==22.1
django-cors-headers==3.13.0
et-xmlfile==1.1.0
fonttools @ file:///D:/bld/fonttools_1651018175663/work
gunicorn==20.1.0
h11==0.13.0
html-table-parser==0.1.0
idna==3.3
kiwisolver @ file:///D:/bld/kiwisolver_1655141740116/work
matplotlib @ file:///D:/bld/matplotlib-suite_1651609712471/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
munkres==1.1.4
mysqlclient==2.1.1
numpy==1.23.0
openpyxl==3.0.10
outcome==1.2.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.3
Pillow==9.1.1
plotly==5.9.0
psycopg2-binary==2.9.3
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
schedule==1.1.0
scipy==1.8.1
seaborn==0.11.2
selenium==4.2.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tenacity==8.0.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
unicodedata2 @ file:///D:/bld/unicodedata2_1649112137711/work
urllib3==1.26.9
whitenoise==6.2.0
wincertstore==0.2
wordcloud @ file:///D:/bld/wordcloud_1637007587460/work
wsproto==1.1.0
